Crate ssb_verify_signatures

Verify Secure Scuttlebutt message signatures (in parallel)

How is this different to ssb-legacy-msg?

It’s built on top of ssb-legacy-msg and ssb-legacy-msg-data but exposed a hopefully easier api, and most importantly it lets you batch process a collection of messages.

Batch processing is good for two reasons:

it means we can utilise multiple cores using rayon

it means we can use the ed25519_dalek verify_batch function that takes advantage of processor SIMD instuctions.

Benchmarking on a 2016 2 core i5 shows that batch processing with par_verify_messages is ~3.6 times faster than using verify_message

Benchmarking on Android on a One Plus 5T (8 core arm64) shows that batch processing with par_verify_messages is ~9.9 times faster than using verify_message!

Enums

Error

Constants

CHUNK_SIZE

Functions

par_verify_message_values
Checks signatures of a slice of message values in parallel.

par_verify_messages
Checks signatures of a slice of messages in parallel.

verify_message
Verify the signature of an entire ssb message that has key and value.

verify_message_value
Verify the signature of a ssb message.value.